release

+++ src/pages/Camera/index.tsx
@@ -1,4 +1,4 @@
-import { Box, Paper } from "@mui/material";
+import { Box, Paper, Stack } from "@mui/material";
 import useCamera from "queries/useCamera";
 import targetImage from "assets/target.jpg";
 import targetTest from "assets/target_test.jpg";
@@ -6,9 +6,7 @@ import { useNavigate, useParams } from "react-router-dom";
 import { useUpdateEffect } from "usehooks-ts";
 import { useEffect, useRef } from "react";
 import cv from "@techstark/opencv-js";
-import { clearInterval } from "timers";
 import Webcam from "react-webcam";
-import { display } from "@mui/system";
 
 export default function Camera() {
   const { id } = useParams<{ id: string }>();
@@ -17,6 +15,7 @@ export default function Camera() {
   const canvasRef = useRef(null);
   const imageRef = useRef(null);
   const realImageRef = useRef<HTMLImageElement>(null);
+  const image2Ref = useRef<HTMLCanvasElement>(null);
   const webcamRef = useRef<Webcam>(null);
   useUpdateEffect(() => {
     if (!data?.some(({ deviceId }) => deviceId === id)) {
@@ -33,37 +32,61 @@ export default function Camera() {
         return new Promise((resolve) => {
           if (realImageRef.current) {
             realImageRef.current.src = imageSrc;
+            // realImageRef.current.src = targetImage;
+
             realImageRef.current.onload = () => {
               try {
-                if (realImageRef.current && canvasRef.current) {
+                if (
+                  realImageRef.current &&
+                  canvasRef.current &&
+                  image2Ref.current
+                ) {
                   //test only
                   const img = cv.imread(realImageRef.current);
 
                   const gray = new cv.Mat();
                   cv.cvtColor(img, gray, cv.COLOR_BGR2GRAY);
-
+                  cv.GaussianBlur(
+                    gray,
+                    gray,
+                    new cv.Size(3, 3),
+                    0,
+                    0,
+                    cv.BORDER_DEFAULT
+                  );
+                  cv.medianBlur(gray, gray, 3);
                   const edges = new cv.Mat();
                   cv.Canny(gray, edges, 50, 150, 3);
-                  // const circles = new cv.Mat();
-                  // cv.HoughCircles(
-                  //   edges,
-                  //   circles,
-                  //   cv.HOUGH_GRADIENT,
-                  //   1,
-                  //   20,
-                  //   50,
-                  //   30,
-                  //   0,
-                  //   0
-                  // );
+                  const circles = new cv.Mat();
+                  cv.HoughCircles(
+                    edges,
+                    circles,
+                    cv.HOUGH_GRADIENT,
+                    1,
+                    50,
+                    175,
+                    30,
+                    0,
+                    0
+                  );
+                  console.log(circles);
+                  let color = new cv.Scalar(100, 100, 100);
+                  for (let i = 0; i < circles.cols; i++) {
+                    let x = circles.data32F[i * 3];
+                    let y = circles.data32F[i * 3 + 1];
+                    let radius = circles.data32F[i * 3 + 2];
+                    let center = new cv.Point(x, y);
+                    cv.circle(img, center, radius, color);
+                  }
 
-                  cv.imshow(canvasRef.current, edges);
+                  cv.imshow(image2Ref.current, edges);
+                  cv.imshow(canvasRef.current, img);
                   img.delete();
                   gray.delete();
+                  circles.delete();
                   resolve(1);
                 }
               } catch (err) {
-                console.log(err);
                 resolve(1);
               }
             };
@@ -74,7 +97,7 @@ export default function Camera() {
 
     let interval = setInterval(async () => {
       await imageProcessor();
-    }, 1000 / 30);
+    }, 1000);
 
     return () => {
       clearInterval(interval);
@@ -86,7 +109,7 @@ export default function Camera() {
         alt="target_test"
         src={targetTest}
         height="480px"
-        width={"480px"}
+        width={"640px"}
         style={{ display: "none" }}
         ref={realImageRef}
       />
@@ -97,18 +120,22 @@ export default function Camera() {
         height={"480px"}
         width={"640px"}
         ref={webcamRef}
+        style={{ display: "none" }}
       />
       <img
         alt="target"
         src={targetImage}
         height="480px"
-        width={"480px"}
+        width={"640px"}
         style={{ display: "none" }}
         ref={imageRef}
       />
-      <Box>
-        <canvas height={"480px"} width="480px" ref={canvasRef} />
-      </Box>
+      <Stack direction={"row"}>
+        <Box>
+          <canvas ref={canvasRef} />
+        </Box>
+        <canvas ref={image2Ref} />
+      </Stack>
     </Paper>
   );
 }
